Fix headerfinder to look exclusively at headers
Previously, if a line started with a header character, but was not
followed by a space, it would still be included. Now, look explicitly
for something matching `<CHAR> <anything>`.

This prevents cases like org-mode, where the header character (*) is
also used for bold text, so bold text at the start of a line was getting
counted as a header.